+++ src/output_plugins/logger_pluginStdout.h
@@ -24,7 +24,7 @@ extern "C" {
 
 LOGGER_STATUS logger_stdout_initialize ( LOGGER_INI_SECTIONHANDLE paramBag );
 LOGGER_STATUS logger_stdout_terminate ( void );
-LOGGER_STATUS logger_stdout_transmit ( char * msg, uint32_t msgLen );
+LOGGER_STATUS logger_stdout_transmit ( char * msg, size_t msgLen );
 char* logger_stdout_name ( void );
