+++ src/output_plugins/logger_pluginUdp.h
@@ -25,7 +25,7 @@ extern "C" {
 
 LOGGER_STATUS logger_udp_initialize ( LOGGER_INI_SECTIONHANDLE paramBag );
 LOGGER_STATUS logger_udp_terminate ( void );
-LOGGER_STATUS logger_udp_transmit ( char * msg, uint32_t msgLen );
+LOGGER_STATUS logger_udp_transmit ( char * msg, size_t msgLen );
 char * logger_udp_name ( void );
